backend/requirements.txt
django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==21.2.0
<<<<<<< HEAD
psycopg2-binary
=======
whitenoise==6.8.2
>>>>>>> 8740a5026e07176ae407719cb11814b7fc1812f6
